Add goconfig files.
Package goconfig provides a wrapper for the koanf/v2 configuration
package that make it easier to use for a preferred use case, configuring
it to load values (in order) from yaml files, environment variables and
flags.

Due to unimplemented functionality in koanf, it is necessary to use the
pflags flag packge, as the merging for the go library version does not
properly deal with flag default values.